more parameters written to the output file, io of old output tested
git-svn-id: https://svn.fysik.dtu.dk/projects/gpaw/trunk@1337 6419b465-d902-0410-997c-be2266baf723

+++ test/lrtddft.py
@@ -5,46 +5,57 @@ from gpaw.utilities import equal
 from ASE import ListOfAtoms, Atom, Crystal
 from gpaw.lrtddft import LrTDDFT
 
-R=0.7 # approx. experimental bond length
-a = 2
-c = 3
-H2 = ListOfAtoms([Atom('H', (a/2,a/2,(c-R)/2)),
-                  Atom('H', (a/2,a/2,(c+R)/2))],
-                 cell=(a,a,c))
-calc = Calculator(xc='PBE',nbands=2,spinpol=False)
-H2.SetCalculator(calc)
-H2.GetPotentialEnergy()
+io_only=False
 
-xc='LDA'
+if not io_only:
+    R=0.7 # approx. experimental bond length
+    a = 2
+    c = 3
+    H2 = ListOfAtoms([Atom('H', (a/2,a/2,(c-R)/2)),
+                      Atom('H', (a/2,a/2,(c+R)/2))],
+                     cell=(a,a,c))
+    calc = Calculator(xc='PBE',nbands=2,spinpol=False)
+    H2.SetCalculator(calc)
+    H2.GetPotentialEnergy()
 
-# without spin
-lr = LrTDDFT(calc,xc=xc)
-lr.diagonalize()
-t1 = lr[0]
+    xc='LDA'
 
-# with spin
-lr = LrTDDFT(calc,xc=xc,nspins=2)
-lr.diagonalize()
-# the triplet is lower, so that the second is the first singlet
-# excited state
-t2 = lr[1]
-
-equal(t1.get_energy(), t2.get_energy(), 5.e-7)
-
-# course grids
-for finegrid in [1,0]:
-    lr = LrTDDFT(calc, xc=xc, finegrid=finegrid)
+    # without spin
+    lr = LrTDDFT(calc,xc=xc)
     lr.diagonalize()
-    t3 = lr[0]
-    print "finegrid, t1, t3=", finegrid, t1.get_energy(),t3.get_energy()
-    equal(t1.get_energy(), t3.get_energy(), 5.e-4)
+    t1 = lr[0]
+
+    # with spin
+    lr = LrTDDFT(calc,xc=xc,nspins=2)
+    lr.diagonalize()
+    # the triplet is lower, so that the second is the first singlet
+    # excited state
+    t2 = lr[1]
+
+    equal(t1.get_energy(), t2.get_energy(), 5.e-7)
+
+    # course grids
+    for finegrid in [1,0]:
+        lr = LrTDDFT(calc, xc=xc, finegrid=finegrid)
+        lr.diagonalize()
+        t3 = lr[0]
+        print "finegrid, t1, t3=", finegrid, t1.get_energy(),t3.get_energy()
+        equal(t1.get_energy(), t3.get_energy(), 5.e-4)
 
 # io
 fname = 'lr.dat.gz'
-lr.write(fname)
+if not io_only:
+    lr.write(fname)
 lr = LrTDDFT(filename=fname)
-os.remove(fname)
+if not io_only:
+    os.remove(fname)
 t4 = lr[0]
 
-equal(t3.get_energy(), t4.get_energy(), 1.e-6)
+if not io_only:
+    equal(t3.get_energy(), t4.get_energy(), 1.e-6)
 
+# io of old output
+
+lr = LrTDDFT(filename='lr_old.dat.gz')
+t5 = lr[0]
+equal(t4.get_energy(), t5.get_energy(), 1.e-12)
